feat(dlight): widen the dlight mask to 64 bits and double MAX_DLIGHTS
surf->dlightbits was a plain int and R_MarkLights took bit = 1<<i, so
MAX_DLIGHTS=32 exactly saturated the word.  Raising the pool without widening
that first would have been silent corruption rather than a bigger pool, which
is why the eviction work (1c1864070, 4f11ad9ec) could only mitigate
uhexen2-ck6h instead of curing it.

Widened to unsigned long long across all three copies of msurface_t
(render.h, model.h, gl_model.h), R_MarkLights now takes an unsigned long long
bit, and every call site passes 1ULL<<k.  MAX_DLIGHTS 32 -> 64.

Against the measured peak CL_AllocDlight demand -- meso 74/frame, palace 32,
keep 26 -- 64 clears palace and keep outright and cuts meso's shortfall from
42 to 10.  It does NOT fully cure meso; 64 is the ceiling for a single-word
mask, and going past it needs a multi-word mask or an R_MarkLights that takes
the light index and does word/bit itself.  uhexen2-liqz carries that.

Also bounded the two dlightbits-gated per-surface scans by the highest set bit
instead of always walking MAX_DLIGHTS (gl_rsurf.c, r_surf.c).  Doubling the
pool would otherwise have doubled that scan for every lit surface; this makes
it proportional to the lights actually touching the surface, which is a win at
32 as well.  The lnum < MAX_DLIGHTS term is deliberately first in the
condition -- shifting a 64-bit value by 64 is undefined.

Verified: builds clean, and demo1, egypt1, sot/meso and sot/palace all reach
the first world draw.

NOT verified, and worth knowing: nothing here exercised an index >= 32 at
runtime.  Every headless run showed zero live dlights at spawn, because the
client disconnects a few frames in under this harness -- too little simulation
time for torch thinks, not a contradiction of the 32/32 figure measured on
meso previously.  The >32 half of the mask is therefore correct by
construction and by compile, not by observation.

The software renderer shares r_light.c and r_surf.c and is built only by the
WASM software target, which needs a toolchain this box does not have, so those
three edits are unbuilt.  They are line-for-line the same widening as the GL
side that does build and run.

Refs: uhexen2-liqz, uhexen2-ck6h

+++ engine/hexen2/gl_rmain.c
@@ -2550,7 +2550,7 @@ static qboolean		world_inst_collected[MAX_VISEDICTS];
 static qboolean		world_inst_needs_legacy[MAX_VISEDICTS];
 
 extern qboolean		R_CullBox (vec3_t mins, vec3_t maxs);
-extern void		R_MarkLights (dlight_t *light, int bit, mnode_t *node);
+extern void		R_MarkLights (dlight_t *light, unsigned long long bit, mnode_t *node);
 extern int		c_brush_polys;
 extern GLuint		world_vao;
 extern GLuint		world_ibo;
@@ -2713,12 +2713,12 @@ void R_CollectBrushInstances (void)
 					tdl.origin[0] =  DotProduct (tmp, fwd);
 					tdl.origin[1] = -DotProduct (tmp, rt);
 					tdl.origin[2] =  DotProduct (tmp, up);
-					R_MarkLights (&tdl, 1<<k,
+					R_MarkLights (&tdl, 1ULL<<k,
 					    clmodel->nodes + clmodel->hulls[0].firstclipnode);
 				}
 				else
 				{
-					R_MarkLights (dl, 1<<k,
+					R_MarkLights (dl, 1ULL<<k,
 					    clmodel->nodes + clmodel->hulls[0].firstclipnode);
 				}
 			}
